# packages/schemas/openapi.json
{
  "components": {
    "schemas": {
...
      "BudgetResponse": {
        "properties": {
          "cap_display": {
            "title": "Cap Display",
            "type": "string"
          },
          "cap_micros": {
            "title": "Cap Micros",
            "type": "integer"
          },
          "percent_used": {
            "title": "Percent Used",
            "type": "integer"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "space_spent_display": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Spent Display"
          },
          "space_spent_micros": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Spent Micros"
          },
          "spent_display": {
            "title": "Spent Display",
            "type": "string"
          },
          "spent_micros": {
            "title": "Spent Micros",
            "type": "integer"
          }
        },
        "required": [
          "period",
          "spent_micros",
          "cap_micros",
          "percent_used",
          "spent_display",
          "cap_display"
        ],
        "title": "BudgetResponse",
        "type": "object"
      },
      "ChannelIdentity": {
        "additionalProperties": false,
        "description": "One allowlist entry: this person, on this channel, is that identity.",
        "properties": {
          "channel": {
            "const": "discord",
            "title": "Channel",
            "type": "string"
          },
          "external_user_id": {
            "minLength": 1,
            "title": "External User Id",
            "type": "string"
          },
          "identity": {
            "minLength": 1,
            "title": "Identity",
            "type": "string"
          }
        },
        "required": [
          "channel",
          "external_user_id",
          "identity"
        ],
        "title": "ChannelIdentity",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RiskLevel": {
        "description": "How much damage a tool call can do. Policy is a *set* of levels, never a threshold.",
        "enum": [
          "low",
          "medium",
          "high"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "SettingsResponse": {
        "description": "Workspace settings plus the read-only facts the UI needs beside them.",
        "properties": {
          "configured_secrets": {
            "items": {
              "type": "string"
            },
            "title": "Configured Secrets",
            "type": "array"
          },
          "data_dir": {
            "title": "Data Dir",
            "type": "string"
          },
          "known_secrets": {
            "items": {
              "type": "string"
            },
            "title": "Known Secrets",
            "type": "array"
          },
          "model_is_priced": {
            "title": "Model Is Priced",
            "type": "boolean"
          },
          "settings": {
            "$ref": "#/components/schemas/WorkspaceSettings"
          },
          "supported_providers": {
            "items": {
              "type": "string"
            },
            "title": "Supported Providers",
            "type": "array"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "settings",
          "configured_secrets",
          "known_secrets",
          "supported_providers",
          "model_is_priced",
          "version",
          "data_dir"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
...
      "ToolPolicy": {
        "description": "What the gate does with a call to one tool, by name, before its risk level is read.\n\n``ASK`` is the default and means the risk-level rule decides; ``ALLOW``\nruns the call without asking; ``DENY`` refuses it without asking. A tool\nabsent from a policy is ``ASK``.",
        "enum": [
          "ask",
          "allow",
          "deny"
        ],
        "title": "ToolPolicy",
        "type": "string"
      },
...
      "UpdateSettingsRequest": {
        "additionalProperties": false,
        "description": "A partial update. Every field optional; omitted fields are untouched.\n\nUnknown fields are rejected, not dropped: Pydantic's default turns a\nmisspelled setting into a `200 OK` that changed nothing. This model must\nlist every field of :class:`~agentbase.store.settings.WorkspaceSettings`\n(they differ in bounds and optionality, so it cannot be the same class),\nand `test_every_workspace_setting_can_be_patched` keeps the two in step.",
        "properties": {
          "auto_approve": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/RiskLevel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Approve"
          },
          "channel_approvals": {
            "anyOf": [
              {
                "enum": [
                  "dashboard_only",
                  "originator"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Approvals"
          },
          "channel_identities": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ChannelIdentity"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Identities"
          },
          "channel_space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Space Id"
          },
          "discord_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Discord Enabled"
          },
          "max_agents_per_run": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Agents Per Run"
          },
          "max_run_cost_micros": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Run Cost Micros"
          },
          "max_run_seconds": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Run Seconds"
          },
          "max_steps_per_agent": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Steps Per Agent"
          },
          "model": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "monthly_cap_micros": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Cap Micros"
          },
          "ollama_base_url": {
            "anyOf": [
              {
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ollama Base Url"
          },
          "onboarding_completed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onboarding Completed"
          },
          "openai_access": {
            "anyOf": [
              {
                "enum": [
                  "api_key",
                  "chatgpt"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Openai Access"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "tool_policies": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/ToolPolicy"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Policies"
          }
        },
        "title": "UpdateSettingsRequest",
        "type": "object"
      },
...
      "UsageReport": {
        "properties": {
          "by_day": {
            "items": {
              "$ref": "#/components/schemas/UsageBucket"
            },
            "title": "By Day",
            "type": "array"
          },
          "by_model": {
            "items": {
              "$ref": "#/components/schemas/UsageBucket"
            },
            "title": "By Model",
            "type": "array"
          },
          "by_space": {
            "items": {
              "$ref": "#/components/schemas/UsageBucket"
            },
            "title": "By Space",
            "type": "array"
          },
          "cap_micros": {
            "title": "Cap Micros",
            "type": "integer"
          },
          "period": {
            "title": "Period",
            "type": "string"
          },
          "periods": {
            "items": {
              "type": "string"
            },
            "title": "Periods",
            "type": "array"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/UsageRun"
            },
            "title": "Runs",
            "type": "array"
          },
          "space_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space Id"
          },
          "totals": {
            "$ref": "#/components/schemas/UsageTotals"
          }
        },
        "required": [
          "period",
          "periods",
          "space_id",
          "cap_micros",
          "totals",
          "by_model",
          "by_space",
          "by_day",
          "runs"
        ],
        "title": "UsageReport",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerifyResponse": {
        "description": "Whether the current settings can build a provider, and why not if not.",
        "properties": {
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "ok"
        ],
        "title": "VerifyResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AgentBase sidecar",
    "version": "0.4.5"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/budget": {
      "get": {
        "description": "Month-to-date spend against the cap; ``space_id`` adds that space's share beside it.",
        "operationId": "get_budget_budget_get",
        "parameters": [
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Budget"
      }
    },
...
    "/settings": {
      "get": {
        "operationId": "get_settings_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Settings"
      },
      "patch": {
        "description": "Apply a partial settings update.\n\nValidation failures are 400s carrying `{message, field}`.",
        "operationId": "update_settings_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings"
      }
    },
...
    "/settings/verify": {
      "post": {
        "description": "Check the current settings can build a provider, without calling the model.\n\nThe dashboard's pre-flight: the refusal a run would get, shown before\nStart. With ``space_id`` the space's effective settings are checked.",
        "operationId": "verify_provider_settings_verify_post",
        "parameters": [
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Provider"
      }
    },
...
    "/usage": {
      "get": {
        "description": "A month of model calls from the ledger: totals, by model, space and day, and the costliest runs.\n\n``period`` is ``YYYY-MM`` and defaults to this month; ``space_id`` narrows\nto one space's runs. Read from the same rows the cap is enforced on, so the\nreport and the meter cannot disagree.",
        "operationId": "get_usage_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "in": "query",
            "name": "space_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Space Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Usage"
      }
    }
  }
}